linreg3 => linreg, removed linreg, linreg_multi_pheno.
