+++ src/main/scala/is/hail/expr/Type.scala
@@ -282,7 +282,9 @@ case object TFloat extends TNumeric {
   override def genNonmissingValue: Gen[Annotation] = arbitrary[Double].map(_.toFloat)
 
   override def valuesSimilar(a1: Annotation, a2: Annotation, tolerance: Double): Boolean =
-    a1 == a2 || (a1 != null && a2 != null && D_==(a1.asInstanceOf[Float], a2.asInstanceOf[Float], tolerance))
+    a1 == a2 || (a1 != null && a2 != null &&
+      (D_==(a1.asInstanceOf[Float], a2.asInstanceOf[Float], tolerance) ||
+        (a1.asInstanceOf[Double].isNaN && a2.asInstanceOf[Double].isNaN)))
 
   override def scalaClassTag: ClassTag[java.lang.Float] = classTag[java.lang.Float]
 
@@ -303,7 +305,9 @@ case object TDouble extends TNumeric {
   override def genNonmissingValue: Gen[Annotation] = arbitrary[Double]
 
   override def valuesSimilar(a1: Annotation, a2: Annotation, tolerance: Double): Boolean =
-    a1 == a2 || (a1 != null && a2 != null && D_==(a1.asInstanceOf[Double], a2.asInstanceOf[Double], tolerance))
+    a1 == a2 || (a1 != null && a2 != null &&
+      (D_==(a1.asInstanceOf[Double], a2.asInstanceOf[Double], tolerance) ||
+        (a1.asInstanceOf[Double].isNaN && a2.asInstanceOf[Double].isNaN)))
 
   override def scalaClassTag: ClassTag[java.lang.Double] = classTag[java.lang.Double]
