+++ src/main/scala/is/hail/stats/LinearRegressionModel.scala
@@ -2,19 +2,26 @@ package is.hail.stats
 
 import breeze.linalg.{Matrix, Vector}
 import is.hail.annotations.Annotation
+import is.hail.expr.{TArray, TDouble, TInt, TStruct}
 import net.sourceforge.jdistlib.T
 
 object LinearRegressionModel {
-    def fit(x: Vector[Double], y: Vector[Double], yyp: Double, qt: Matrix[Double], qty: Vector[Double], d: Int): Annotation = {
-      val qtx = qt * x
-      val xxp = (x dot x) - (qtx dot qtx)
-      val xyp = (x dot y) - (qtx dot qty)
+  def schema = TStruct(
+    ("beta", TDouble),
+    ("se", TDouble),
+    ("tstat", TDouble),
+    ("pval", TDouble))
 
-      val b = xyp / xxp
-      val se = math.sqrt((yyp / xxp - b * b) / d)
-      val t = b / se
-      val p = 2 * T.cumulative(-math.abs(t), d, true, false)
+  def fit(x: Vector[Double], y: Vector[Double], yyp: Double, qt: Matrix[Double], qty: Vector[Double], d: Int): Annotation = {
+    val qtx = qt * x
+    val xxp = (x dot x) - (qtx dot qtx)
+    val xyp = (x dot y) - (qtx dot qty)
 
-      Annotation(b, se, t, p)
-    }
+    val b = xyp / xxp
+    val se = math.sqrt((yyp / xxp - b * b) / d)
+    val t = b / se
+    val p = 2 * T.cumulative(-math.abs(t), d, true, false)
+
+    Annotation(b, se, t, p)
+  }
 }
