+++ src/main/scala/is/hail/variant/VariantDataset.scala
@@ -448,25 +448,15 @@ class VariantDatasetFunctions(private val vds: VariantDataset) extends AnyVal {
     LDPrune(vds, r2Threshold, windowSize, nCores, memoryPerCore * 1024L * 1024L)
   }
 
-  def linreg(y: String, covariates: Array[String] = Array.empty[String], root: String = "va.linreg", useDosages: Boolean = false, minAC: Int = 1, minAF: Double = 0d): VariantDataset = {
-    requireSplit("linear regression")
-    LinearRegression(vds, y, covariates, root, useDosages, minAC, minAF)
-  }
-
   def linregBurden(keyName: String, variantKeys: String, singleKey: Boolean, aggExpr: String, y: String, covariates: Array[String] = Array.empty[String]): (KeyTable, KeyTable) = {
     requireSplit("linear burden regression")
     vds.requireSampleTString("linear burden regression")
     LinearRegressionBurden(vds, keyName, variantKeys, singleKey, aggExpr, y, covariates)
   }
 
-  def linregMultiPheno(ys: Array[String], covariates: Array[String] = Array.empty[String], root: String = "va.linreg", useDosages: Boolean = false, minAC: Int = 1, minAF: Double = 0d): VariantDataset = {
-    requireSplit("linear regression for multiple phenotypes")
-    LinearRegressionMultiPheno(vds, ys, covariates, root, useDosages, minAC, minAF)
-  }
-
-  def linreg3(ys: Array[String], covariates: Array[String] = Array.empty[String], root: String = "va.linreg", useDosages: Boolean = false, variantBlockSize: Int = 16): VariantDataset = {
+  def linreg(ys: Array[String], covariates: Array[String] = Array.empty[String], root: String = "va.linreg", useDosages: Boolean = false, variantBlockSize: Int = 16): VariantDataset = {
     requireSplit("linear regression 3")
-    LinearRegression3(vds, ys, covariates, root, useDosages, variantBlockSize)
+    LinearRegression(vds, ys, covariates, root, useDosages, variantBlockSize)
   }
 
   def lmmreg(kinshipMatrix: KinshipMatrix,
